Fix: Multiple compilation errors
- Add lifecycle-runtime-compose dependency for collectAsStateWithLifecycle
- Fix duplicate companion object in DnsVpnService
- Fix shr/and bitwise operators (Short→Int)
- Fix isRunning→isBenchmarkRunning property name
- Fix suspend function call in DnsRepository init
- Fix Converters KSP type resolution

+++ app/src/main/java/com/dnsoptimizer/pro/domain/benchmark/DnsBenchmarkEngine.kt
@@ -509,7 +509,7 @@ class DnsBenchmarkEngine {
         val question = mutableListOf<Byte>()
         
         // Transaction ID
-        val transactionId = (Math.random() * 65535).toInt().toShort()
+        val transactionId = (Math.random() * 65535).toInt()
         question.addAll(listOf(
             (transactionId shr 8).toByte(),
             (transactionId and 0xFF).toByte()
